requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
pyzbar==0.1.9
qrcode==8.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.13.0
uvicorn==0.34.0
